Add nested macros for generic STANAG packet construction
