--- core/klv.cpp
@@ -1,112 +1,0 @@
-#include "klv.h"
-#include <algorithm>
-#include <stdexcept>
-
-KLVLeaf::KLVLeaf(const UL& ul, double value) : ul_(ul), value_(value) {}
-
-std::vector<uint8_t> KLVLeaf::encode() const {
-    auto* entry = KLVRegistry::instance().find(ul_);
-    if (!entry) throw std::runtime_error("Unknown UL");
-    auto data = entry->encoder(value_);
-    std::vector<uint8_t> out;
-    out.insert(out.end(), ul_.begin(), ul_.end());
-    out.push_back(static_cast<uint8_t>((data.size() >> 8) & 0xFF));
-    out.push_back(static_cast<uint8_t>(data.size() & 0xFF));
-    out.insert(out.end(), data.begin(), data.end());
-    return out;
-}
-
-void KLVLeaf::decode(const std::vector<uint8_t>& bytes) {
-    if (bytes.size() < 18) throw std::runtime_error("Too short");
-    UL ul;
-    std::copy(bytes.begin(), bytes.begin() + 16, ul.begin());
-    if (ul != ul_) throw std::runtime_error("UL mismatch");
-    size_t len = (static_cast<size_t>(bytes[16]) << 8) | bytes[17];
-    if (bytes.size() < 18 + len) throw std::runtime_error("Length mismatch");
-    std::vector<uint8_t> data(bytes.begin() + 18, bytes.begin() + 18 + len);
-    auto* entry = KLVRegistry::instance().find(ul_);
-    if (!entry) throw std::runtime_error("Unknown UL");
-    value_ = entry->decoder(data);
-}
-
-KLVBytes::KLVBytes(const UL& ul, const std::vector<uint8_t>& value)
-    : ul_(ul), value_(value) {}
-
-std::vector<uint8_t> KLVBytes::encode() const {
-    std::vector<uint8_t> out;
-    out.insert(out.end(), ul_.begin(), ul_.end());
-    out.push_back(static_cast<uint8_t>((value_.size() >> 8) & 0xFF));
-    out.push_back(static_cast<uint8_t>(value_.size() & 0xFF));
-    out.insert(out.end(), value_.begin(), value_.end());
-    return out;
-}
-
-void KLVBytes::decode(const std::vector<uint8_t>& bytes) {
-    if (bytes.size() < 18) throw std::runtime_error("Too short");
-    UL ul;
-    std::copy(bytes.begin(), bytes.begin() + 16, ul.begin());
-    if (ul != ul_) throw std::runtime_error("UL mismatch");
-    size_t len = (static_cast<size_t>(bytes[16]) << 8) | bytes[17];
-    if (bytes.size() < 18 + len) throw std::runtime_error("Length mismatch");
-    value_.assign(bytes.begin() + 18, bytes.begin() + 18 + len);
-}
-
-void KLVSet::add(std::shared_ptr<KLVNode> node) {
-    children_.push_back(node);
-}
-
-std::vector<uint8_t> KLVSet::encode() const {
-    std::vector<uint8_t> out;
-    for (const auto& child : children_) {
-        auto data = child->encode();
-        out.insert(out.end(), data.begin(), data.end());
-    }
-    return out;
-}
-
-void KLVSet::decode(const std::vector<uint8_t>& data) {
-    children_.clear();
-    size_t i = 0;
-    while (i + 18 <= data.size()) {
-        UL ul;
-        std::copy(data.begin() + i, data.begin() + i + 16, ul.begin());
-        i += 16;
-        if (i + 2 > data.size()) break;
-        size_t len = (static_cast<size_t>(data[i]) << 8) | data[i + 1];
-        i += 2;
-        if (i + len > data.size()) break;
-        std::vector<uint8_t> value(data.begin() + i, data.begin() + i + len);
-        i += len;
-
-        const KLVEntry* entry = KLVRegistry::instance().find(ul);
-        if (entry) {
-            std::vector<uint8_t> item;
-            item.insert(item.end(), ul.begin(), ul.end());
-            item.push_back(static_cast<uint8_t>((len >> 8) & 0xFF));
-            item.push_back(static_cast<uint8_t>(len & 0xFF));
-            item.insert(item.end(), value.begin(), value.end());
-            auto leaf = std::make_shared<KLVLeaf>(ul);
-            leaf->decode(item);
-            children_.push_back(leaf);
-        } else {
-            auto bytes = std::make_shared<KLVBytes>(ul, value);
-            children_.push_back(bytes);
-        }
-    }
-}
-
-void KLVRegistry::register_ul(const UL& ul, const KLVEntry& entry) {
-    entries_[ul] = entry;
-}
-
-const KLVEntry* KLVRegistry::find(const UL& ul) const {
-    auto it = entries_.find(ul);
-    if (it != entries_.end()) return &it->second;
-    return nullptr;
-}
-
-KLVRegistry& KLVRegistry::instance() {
-    static KLVRegistry inst;
-    return inst;
-}
-
