+++ core/klv_types.h
@@ -1,0 +1,5 @@
+#pragma once
+#include <array>
+#include <cstdint>
+
+using UL = std::array<uint8_t,16>;
