+++ st0601/st0601.cpp
@@ -5,6 +5,7 @@ namespace st0601 {
 
 // Dummy UL values for demonstration purposes
 const UL PLATFORM_HEADING_ANGLE   = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0x00,0x00,0x00};
+const UL UNIX_TIMESTAMP            = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0xFF,0x00,0x01};
 const UL TARGET_LOCATION_ELEVATION = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0x00,0x00,0x01};
 const UL TARGET_TRACK_GATE_WIDTH   = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0x00,0x00,0x02};
 const UL TARGET_TRACK_GATE_HEIGHT  = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0x00,0x00,0x03};
@@ -126,6 +127,25 @@ const UL WEAPONS_STORES            = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x
 const UL WAYPOINT_LIST             = {0x06,0x0E,0x2B,0x34,0x02,0x0B,0x01,0x01,0x01,0x01,0x01,0x01,0x01,0x00,0x00,0x7E};
 
 void register_st0601(KLVRegistry& reg) {
+    // Unix Timestamp: microseconds since epoch -> uint64
+    reg.register_ul(UNIX_TIMESTAMP, {
+        [](double micros) {
+            uint64_t raw = static_cast<uint64_t>(micros);
+            std::vector<uint8_t> out(8);
+            for (int i = 0; i < 8; ++i) {
+                out[7 - i] = static_cast<uint8_t>(raw >> (i * 8));
+            }
+            return out;
+        },
+        [](const std::vector<uint8_t>& bytes) {
+            if (bytes.size() != 8) return 0.0;
+            uint64_t raw = 0;
+            for (uint8_t b : bytes) {
+                raw = (raw << 8) | b;
+            }
+            return static_cast<double>(raw);
+        }
+    });
     // Platform Heading Angle: degrees [0,360] -> uint16
     reg.register_ul(PLATFORM_HEADING_ANGLE, {
         [](double degrees) {
