+++ st0601/st0601.h
@@ -7,6 +7,7 @@ namespace st0601 {
 
 // Universal Labels for a subset of MISB ST0601 tags
 extern const UL PLATFORM_HEADING_ANGLE;
+extern const UL UNIX_TIMESTAMP;
 extern const UL TARGET_LOCATION_ELEVATION;
 extern const UL TARGET_TRACK_GATE_WIDTH;
 extern const UL TARGET_TRACK_GATE_HEIGHT;
